# openapi.yaml
openapi: 3.0.3
info:
  title: Prompt Drift Gateway API
  description: The centralized microservice for versioning, proxying, and testing LLM prompts. Decoupling stateless business logic from AI execution.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Sandbox Environment
paths:
  /v1/prompts:
    post:
      summary: Create a versioned prompt
      description: Ingests a new prompt, assigns a cryptographic SHA-256 hash, and stores it immutably.
      operationId: createPrompt
      tags:
        - Ingestion
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePromptRequest'
      responses:
        '201':
          description: Prompt successfully created and hashed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptResponse'
        '409':
          description: Conflict. Prompt ID already exists.
  /v1/prompts/execute:
    post:
      summary: Execute a versioned prompt
      description: Injects variables into a mathematically versioned prompt and proxies the request to the designated LLM endpoint.
      operationId: executePrompt
      tags:
        - Execution
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteRequest'
      responses:
        '200':
          description: Successful execution and LLM response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecuteResponse'
        '400':
          description: Bad Request.
        '403':
          description: Forbidden (Cross-tenant access violation).
        '404':
          description: Prompt ID not found.
        '502':
          description: Bad Gateway.
  /metrics:
    get:
      summary: Prometheus Metrics
      description: Exposes real-time telemetry to the Prometheus scraper.
      tags:
        - Observability
      responses:
        '200':
          description: Raw time-series text data.
          content:
            text/plain:
              schema:
                type: string
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    CreatePromptRequest:
      type: object
      required:
        - prompt_id
        - template_text
      properties:
        prompt_id:
          type: string
          example: "customer-support-v1"
        template_text:
          type: string
          example: "You are a support agent. Answer {user_name}'s question."
    PromptResponse:
      type: object
      properties:
        prompt_id:
          type: string
        version_hash:
          type: string
        created_at:
          type: string
          format: date-time
    ExecuteRequest:
      type: object
      required:
        - prompt_id
      properties:
        prompt_id:
          type: string
          example: "customer-support-v1"
        variables:
          type: object
          additionalProperties:
            type: string
          example:
            user_name: "Alice"
        model_override:
          type: string
          example: "llama3.1"
    ExecuteResponse:
      type: object
      properties:
        success:
          type: boolean
        executed_prompt:
          type: string
        llm_response:
          type: string
        telemetry:
          type: object
          properties:
            token_usage:
              type: integer
            latency_ms:
              type: integer
            model_used:
              type: string